Refactor create utility class for get closet
