+++ src/main/java/com/harper/asteroids/ApproachDetector.java
@@ -1,7 +1,5 @@
 package com.harper.asteroids;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.harper.asteroids.model.CloseApproachData;
 import com.harper.asteroids.model.Feed;
 import com.harper.asteroids.model.NearEarthObject;
 
@@ -9,17 +7,11 @@ import com.harper.asteroids.neo.NeoClient;
 import com.harper.asteroids.neo.NeoException;
 import com.harper.asteroids.neo.NeoMapper;
 import java.time.LocalDate;
-import java.util.Comparator;
-import java.util.Date;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
-import javax.ws.rs.client.Client;
-import javax.ws.rs.client.ClientBuilder;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
-import java.io.IOException;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -31,7 +23,6 @@ import java.util.stream.Collectors;
  * Alerts if someone is possibly hazardous.
  */
 public class ApproachDetector {
-    private static final int LIMIT = 10;
     private final NeoClient client = new NeoClient();
     private final NeoMapper mapper = new NeoMapper();
     List<NearEarthObject> neos = new ArrayList<>();
@@ -72,20 +63,6 @@ public class ApproachDetector {
 
         executor.shutdown();
         return neos;
-    }
-
-    public List<CloseApproachData> getClosest(List<NearEarthObject> neos, Date today,
-        Date afterWeek) {
-
-        return neos.stream()
-            .filter(neo -> neo.getCloseApproachData() != null && !neo.getCloseApproachData().isEmpty())
-            .flatMap(neo -> neo.getCloseApproachData().stream())
-            .filter(closest -> closest.getCloseApproachDateTime().after(today)
-                && closest.getCloseApproachDateTime().before(afterWeek))
-            .sorted(Comparator.comparing(
-                CloseApproachData::getCloseApproachDateTime))
-            .limit(LIMIT)
-            .collect(Collectors.toList());
     }
 
 }
